Nesting use-authentication code in a directory
